feat: update payment and checkout models to use decimal fields for amounts and add payment method to payment model

+++ backend/apps/checkout/models.py
@@ -63,7 +63,9 @@ class CheckoutTransaction(models.Model):
         default=0,
     )
 
-    total_amount = models.IntegerField(
+    total_amount = models.DecimalField(
+        max_digits=12,
+        decimal_places=2,
         default=0,
     )
